requirements.txt
asgiref==3.2.3
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0
django-autoslug==1.9.6
django-crispy-forms==1.8.1
django-filter==2.2.0
django-phonenumber-field==4.0.0
django-rest-framework==0.1.0
django-taggit==1.2.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.0
idna==2.8
phonenumbers==8.11.4
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
twilio==6.35.4
urllib3==1.25.8
whitenoise==4.1.4
